requirements.txt
langchain 
openai 
chromadb 
streamlit 
tiktoken 
unstructured 
python-docx 
pdfminer.six
sentence-transformers
langchain-community